+++ docs/static/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added new function `measureAxisLabel` as a preset of `measureText`
 - Added new function `measureLegendLabel` as a preset of `measureText`
 - Created new module `sszvis/measure` with utilities to measure elements
+- Added new function `legendBounds` to compute rows, columns, and widths for color legends
 
 ## 2.0 (Dec 2017)